refactor(ui): CompanionAISettingsController.Dialogs.cs を機能単位 partial に分割
- Dialogs.TargetFilter.cs: BuildTargetFilterSection 系と AITargetSelect/AIRule リスト
  (RebuildTargetSelectList / RebuildTargetRuleList / ShowTargetSelectEditDialog /
   ShowTargetRuleEditDialog / FormatFilterSummary / GetElementLabel /
   CloneTargetSelect / GetSortKeyLabel / FormatTargetSelectSummary /
   FormatTargetRuleSummary)
- Dialogs.Conditions.cs: RebuildConditionChipList / BuildConditionRow /
  BuildConditionInputWidgets / ShowConditionEditorDialog / NormalizeCompareOp /
  FormatConditionSummary / HasFilterConstraints / FormatCompareOpSymbol
- Dialogs.ActionPicker.cs: ShowActionPickerDialog / BuildActionPickerContent /
  Build{Instant,Item,Attack,Sustained,Enum}PickerContent / ShowSustainedDurationDialog /
  IsCasterCategory / ResolveInitialTab / GetPickerTabLabel / PickerTabId enum
- Dialogs.cs に残したもの: ShowModeDetailDialog / 統合行動リスト系
  (RebuildUnifiedActionList / OpenActionRuleEditorForIndex /
   BuildUnifiedActionRow / ShowUnifiedActionRuleEditorDialog 等)、
  ShowModeTransitionDialog、共通ヘルパ (BuildDetailSection / CloneMode /
  CloneRule / CloneTransitionRule / IsActionUnlocked / ResolveActionSlotLabel /
  ResolveInitialTabSafe / EditedRuleMeta struct)
- 各 partial ファイルの先頭コメントで担当範囲を明示
