+++ Assets/MyAsset/Runtime/UI/CompanionAISettingsController.Dialogs.ActionPicker.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 28571fedc6244534ae82b8799a7ba9e1
