+++ Assets/MyAsset/Runtime/UI/CompanionAISettingsController.Dialogs.Conditions.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 99a6de6ddc504db781d3a7510fbb2194
